CORE-7209 Fix deleting while editing bug in grid
Enabling auto-commit on the listStore ensures that when the user
clicks the Delete button, any changes they've made inside the grid
are committed first.  This way, when the listStore looks for the
model the user wishes to delete, the key (the string that was entered)
is up-to-date with the listStore.