CORE-7209 Add overwrite-public flag check
If a tool is in use by apps that have already been
made public, the overwrite-public flag can be used
to proceed with an update to that tool.  This adds a
window which presents the user with the DangerZone
message about overwriting tools used by public
apps and a list of the apps currently using that
tool. If the user tries to delete a tool in use
by public apps, they get a notification letting
them know this is not permitted.