CORE-8321 Fix AppEditor window hiding after moving
The setMaximized(true) was being called during onShow() instead of show(), which meant even if the user tried to minimize the window, the window would stay maximized.
Additionally, AppEditerWindow was overriding the onShow() method without calling super.onShow(). This  was leading to weird element positions which effectively hid the window by positioning it far outside the dimensions of the browser.