CORE-8872 Fix Class Cast Exception being thrown during DND within HT Files
The DND logic within PathListViewer assumed the user would only ever DND files from the Data window (with type DiskResource) to the HT File viewer (with type Splittable).  In the code, there was logic casting the DND source to DiskResource. This is not always the case as users may want to rearrange files within the HT file viewer, which was unaccounted for.