CORE-5032 Fix refresh error when restoring subfolder in Trash
The logic was incorrectly assuming that any time the Select All box was checked that the user must be in the top level Trash directory.  In the case where the user tries to restore a file or folder within a subdirectory in Trash that is also the only item displayed in the grid view, the grid automatically checks the Select All box and the DE restores all items in the Trash folder.  From there, the DE then tries to refresh the subdirectory in Trash which no longer exists because it has been restored.