CORE-9157 Check for null Connection and Channel objects in AMQP logic, close the websocket if the Channel is still null
If AMQP is not responding at all, there will still be NPEs produced.  The point is that we're not storing null as the Connection or Channel object and then never trying to re-establish those connections or channels again.
Closing the Channel in the server side code, causes the DesktopPresenterImpl.initNotificationWebSocket.onClose client code to attempt to reopen the websocket which allows for multiple retries.